Fixed compiler warning on Windows where an int was not explicitly casted to a float.
git-svn-id: https://svn.opendtect.org/od/trunk@29231 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/uiTools/uizrangeinput.h
@@ -38,7 +38,7 @@ template <class T> inline
 void uiZRangeInput::setZRange( T range )
 {
     if ( !isdepth_ && !range.isUdf() )
-	range.scale( ZDomain::Time().userFactor() );
+	range.scale( (float) ZDomain::Time().userFactor() );
     
     setValue( range );
 }
